criação de nova pagina html, ajustes no css, ciração da pagina de inicio do admin, manuseio do javascript e outros ajustes

+++ src/main/java/br/com/ingressa/repository/AdminRepository.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
     Optional<Admin> findByCpf(String cpf);
 }
+
+
